Set character encoding in windows CI (Qiskit/qiskit-terra#5988)
* Set character encoding in windows CI

This commit is an attempt to fix the CI configuration around character
encoding in windows when running tests. The test jobs all rely on stestr
for running the tests which spawns multiple worker processes to execute
the tests using python's stdlib subprocess.Popen. This however can
cause issues with the default character encoding on windows (as reported
in mtreinish/stestr#295) where either the worker processes or the pipe
used for interprocess communication on windows does not carry over the
same character encoding as the parent process and reverts to the default
windows encoding of cp1252. This commit tries setting the character
encoding both through the change code page command and through setting
the LANG env variable.

For testing purposes a test is added to purposefully emit utf8
characters to stdout to verify if this works or not. This can be removed
prior to merging (although it doesn't do any harm by itself).

* DNM: Test things fail today

* Revert "DNM: Test things fail today"

This reverts commit 73c3d5f5b2c9b10af49706bbdb96eeec21bef6ec.

* Re-enable logging on algorithms/opflow windows unit tests

* Fix lint

Co-authored-by: Manoel Marques <manoel.marques@ibm.com>
Co-authored-by: Luciano Bello <bel@zurich.ibm.com>