Correctly require is_good_state on AmplificationProblem constructor (Qiskit/qiskit-terra#6417)
* Correctly require is_good_state on AmplificationProblem constructor

This commit fixes an issue with the constructor of the
AmplificationProblem class. It previously treated the ``is_good_state``
kwarg as optional, however it is required in most cases. This updates
the constructor to raise if it's not present and also clears up the
documentation. It also adds a default case for when a PhaseOracle
object is used for the oracle, we can leverage the
``evaluate_bitstring`` method provided to evaluate a state during
grover.

Fixes Qiskit/qiskit-terra#6416

* Update qiskit/algorithms/amplitude_amplifiers/amplification_problem.py

Co-authored-by: Julien Gacon <gaconju@gmail.com>

* Fix lint

* Update tests

* Fix lint

Co-authored-by: Julien Gacon <gaconju@gmail.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>