Optimise documentation build (Qiskit/qiskit-terra#6865)
* Clear cruft out of Sphinx' conf.py

Remove redundant options and comments from conf.py in order to make it
clearer what is actually being used.

The last use of QISKIT_DOCS (to forcibly import some deprecated
visualisation functions) was removed in c4104aa.

Almost all our use of autodoc is via autosummary, where we have used
template files to overwrite the defaults from autodoc, so setting
autodoc options in conf.py had very little effect, and was quite
misleading.  Also,
    "inherited-members": None
actually _set_ `:inherited-members:` on all `autoclass` directives,
rather than removed it, which was presumably the original intent.

Unnecessary comments exhorting people the read the Sphinx documentation
are removed, and only a short sentence describing the intent remains.
We hope that if you don't understand Sphinx' configuration, the
documentation is the first place you'd look without a comment telling
you so.  Those comments were generated by sphinx-quickstart, not by us,
anyway.  Removing them makes it easier to read the file.

* Rewrite default autosummary templates

This documents all class attributes on the same page as the classes
themselves.  This removes a huge amount of Sphinx stubs, all of which
needed to be processed and rendered into HTML, and mostly only had one
line of text on them anyway.

It also adds a new template
    class_no_inherited_members.rst
that may be used when we do not want to individually document all
function methods that were derived from a parent class.  This template
is assigned to all classes documented within qiskit.circuit.library and
qiskit.opflow; the sheer number of subclasses in those meant that huge
numbers of inherited methods were being redocumented.  This was making
it much harder to see what was actually being redefined by those
methods, and was affecting the build times.  It was particularly slow
for derivatives of QuantumCircuit, since there were several calls to
    .. jupyter-execute ::
directives that were being repeated for every subclass.

This removes the redundant `base.rst` template, which was just the
default behaviour anyway.

The build times should now have shrunk massively.  On my laptop (Macbook
Pro 2020 i7), a clean build on a single core now takes about 12 minutes
(down from 47), and this reduces to 4 minutes when using 4 processes.

* Fix rst heading levels in opflow and algorithms

This removes the unintended extra entries from the API documentation
list for Terra of submodules of opflow and algorithms.  It also converts
various summaries in opflow and algorithms to use autosummary, rather
than an awkward toctree statement that necessitated extra stubs inside
`docs/apidocs`.

We have to re-write some of the titles in submodules of opflow to work
around behaviour of Sphinx' autosummary; when it tries to generate the
one-line summary of the objects in its table, it parses the docstring as
RST, then attempts to summarize the first paragraph.  If the first
syntax element is _not_ a paragraph (e.g. if it is a section heading),
then autosummary simply takes the first non-blank line of the docstring.
This means that if a submodule begins with a top-level RST header that
looks like
    =====
    Title
    =====
then the `=====` will be taken literally, which will cause the rest of
the RST document to fail to parse.

To work around it, we just lower all the heading levels, so that the
highest level heading in submodules is denoted by
    Title
    =====
and subheadings are
    Section Heading
    +++++++++++++++
and so on.

* Standardise heading level markers

* Move attributes after methods

Also stop `__init__` from being generated as a stub; it's automatically
included in class documentation anyway.

* Import missing theme files from metapackage

The files
    docs/_templates/layout.html
    docs/_templates/versions.html
    docs/_templates/theme_variables.jinja
are directly copied in from the metapackage at the current state,
Qiskit/qiskit@24d3df2d3.

* Remove unnecessary list

* Use list not generator in non-empty condition