Revert "Implemented general Quantum Time Evolution Framework interfaces. (Qiskit/qiskit-terra#7669)" (Qiskit/qiskit-terra#7845)
* Revert "Implemented general Quantum Time Evolution Framework interfaces. (Qiskit/qiskit-terra#7669)"

This reverts commit 6566c43f8620c9919cdeb08483d8e0e8f8b4ba56.

* Fix dangling import

After Qiskit/qiskit-terra#7669 the custom ListOrDict type it added was used in Qiskit/qiskit-terra#7758. This
was missed in the straight revert of Qiskit/qiskit-terra#7669 which was causing an import
error. This commit fixes this oversight by updating the the import in
aux_ops_evaluator to use the ListOrDict type where it lives with Qiskit/qiskit-terra#7669
reverted (this is why we should try to keep each PR to a single logical
change and avoid unrelated refactors in a PR).

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>