Prepare for removal of `tweedledum` as a requirement (Qiskit/qiskit-terra#8738)
* Prepare for removal of `tweedledum` as a requirement

This adds the necessary optional-functionality checker
(`HAS_TWEEDLEDUM`), converts all Terra uses of `tweedledum`-dependent
functionality to be gated on lazy checker, and deprecates the implicit
import of `tweedledum`-specific functionality in `qiskit.circuit`.  This
commit does not remove `tweedledum` as a requirement just yet; we are
waiting one Terra release to give downstream users a chance to adjust.

* Fix lint

* Fix import paths

* Fix straggler release note

* Update releasenotes/notes/begin-tweedledum-removal-25bb68fc72804f00.yaml

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>