"sampler" is currently a keyword argument
The example provided specifies the sampler option as if it would be positional, but it is actually a keyword argument.
I don't understand the "*" or the possibility of the sampler to be None. But in order to make it work without removing those, the sampler hast to be especified as "sampler"=sampler.